remove schema-gen
Too closely interop with backend crates results in a terrible build
procses, and frankly we should be replying on our OpenAPI schema to
build the schemas for API endpoints agnostic of language choice. As such
should be replaced with a tool that can handle reading OpenAPI schemas
and generating types from it.