standardize model structs
Generally each model file should correspond directly to the database
object. It should contain a Response alias or struct that may be a
truncated version of data that is exposed to the user. See
`./backend/auth/src/models/user.rs` for an example that does not expose
`password_hash` as part of the Response. It may optionally contain a New
struct that instructs how this row is built in the database.