Add products post endpoint
Also update the client libraries as needed. The new endpoint can support
taking multiple products as well as multiple measure values as one json
object. There is however a fault when attempting to input a value which
already exists in the DB, in that it will not update the existing row.
TODO decide on behavior. It should probably update, given that we will
call this endpoint en-masse for scraper updates.